@@ -14,17 +14,16 @@
  */
 buildscript {
     repositories {
-        jcenter()
         mavenCentral()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:2.3.3'
+        classpath 'com.android.tools.build:gradle:4.2.0'
     }
 }
 
 allprojects {
     repositories {
-        jcenter()
+        mavenCentral()
         maven {
             url "https://maven.google.com"
         }
@@ -44,11 +43,11 @@ if (project.hasProperty("pluginVersion")) {
 project.ext.baseName = "google-signin-cpp"
 
 android {
-    compileSdkVersion 26
-    buildToolsVersion "25.0.3"
+    compileSdkVersion 31
+    buildToolsVersion "30.0.3"
     defaultConfig {
         minSdkVersion 14
-        targetSdkVersion 26
+        targetSdkVersion 31
         versionName project.version
         archivesBaseName = project.ext.baseName
         externalNativeBuild {
@@ -58,7 +57,7 @@ android {
             }
         }
         ndk {
-            abiFilters 'x86', 'armeabi-v7a'
+            abiFilters 'x86', 'x86_64', 'armeabi-v7a', 'arm64-v8a'
         }
     }
     lintOptions {
@@ -117,5 +116,5 @@ project.afterEvaluate {
 }
 
 dependencies {
-    compile 'com.google.android.gms:play-services-auth:11.8.0'
+    compile 'com.google.android.gms:play-services-auth:20.2.0'
 }
